add gamestate and it's driver

+++ modules/InventoryHolder/inventoryholder.hpp
@@ -12,16 +12,16 @@ using namespace std;
 class InventoryHolder{
     protected:
     vector<GameCard*> cardbuffer;
-    const uint8_t num_of_cards;
+    const int num_of_cards;
 
     public:
 
     InventoryHolder(); //ctor
-    InventoryHolder(uint8_t num); //param ctor
+    InventoryHolder(int num); //param ctor
     ~InventoryHolder(); //dtor
 
     vector<GameCard*> get_buffer();
-    virtual uint8_t get_num_of_cards() const = 0;
+    virtual int get_num_of_cards() const = 0;
 
 };
